requirements.txt
backports.tarfile==1.2.0
boto3==1.35.0
fqdn==1.5.1
httptools==0.6.1
inflect==7.3.1
isoduration==20.11.0
jaraco.collections==5.1.0
langchain-chroma==0.1.4
langchain-experimental==0.3.0
langchain-openai==0.2.0
loadenv==0.1.1
mito-ai==0.1.1
pandas-datareader==0.10.0
pip-chill==1.0.3
plotly==5.24.1
pygwalker==0.4.9.10
pypdf==5.0.0
seaborn==0.13.2
sentence-transformers==3.1.1
streamlit==1.39.0
textblob==0.18.0.post0
tomli==2.0.1
uri-template==1.3.0
watchfiles==0.24.0
webcolors==24.8.0
wordcloud==1.9.3
xgboost==2.1.1
xlrd==2.0.1
yahoo-fin==0.8.9.1
yfinance==0.2.44
